Add CoA agent baseline
This adds the CoA agent into the list of baseline flows which will run
during study seeding, if the agent is part of the search space. This is
not necessary, but helps ensure Optuna sees an instance of the agent
before beginning the optimization phase.

We also add the agent to the Flows enum and the get_flow_name function

Finally, we ran a test study with the following study config:

```yaml
---
name: "coa-rag-agent-test-a100"
dataset:
  xname: "financebench_hf"

reuse_study: true

optimization:
  cpus_per_trial: 1
  gpus_per_trial: 0.2
  max_concurrent_trials: 25
  num_eval_samples: 100
  num_retries_unique_params: 10
  num_trials: 100

search_space:
  rag_modes:
  - "coa_rag_agent"
```